Testing and restructuring for test
